Refactor: 메인 액티비티/Decorator 분리
