+++ app/src/main/java/com/diary/paintlog/view/activities/MainActivity.kt
@@ -1,33 +1,18 @@
 package com.diary.paintlog.view.activities
 
-import android.content.Context
-import android.graphics.Color
-import android.os.Build
+import android.content.Intent
 import android.os.Bundle
-import android.text.SpannableString
-import android.text.Spanned
-import android.text.style.ForegroundColorSpan
-import android.util.Log
 import android.view.MenuItem
 import android.view.View
-import android.view.WindowInsetsController
-import android.view.WindowManager
-import android.view.animation.AlphaAnimation
 import android.widget.ImageView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.view.ContextThemeWrapper
 import androidx.appcompat.widget.PopupMenu
-import androidx.constraintlayout.widget.Group
-import androidx.core.content.ContextCompat
 import com.diary.paintlog.R
 import com.diary.paintlog.databinding.ActivityMainBinding
+import com.diary.paintlog.utils.decorator.CalendarDecorator
 import com.prolificinteractive.materialcalendarview.CalendarDay
-import com.prolificinteractive.materialcalendarview.DayViewDecorator
-import com.prolificinteractive.materialcalendarview.DayViewFacade
-import com.prolificinteractive.materialcalendarview.format.WeekDayFormatter
-import com.prolificinteractive.materialcalendarview.spans.DotSpan
-import org.threeten.bp.DayOfWeek
 
 class MainActivity : AppCompatActivity() {
 
@@ -43,11 +28,14 @@ class MainActivity : AppCompatActivity() {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
+        // 일자를 꾸며줄 데코레이터 클래스
+        val calendarDecorator = CalendarDecorator(binding.calendarView.context)
+
         // 좌우 화살표 사이 연, 월의 폰트 스타일 설정
         binding.calendarView.setHeaderTextAppearance(R.style.CalendarWidgetHeader)
 
         // 요일을 한글로 변경하고 요일별로 색을 바꿈
-        binding.calendarView.setWeekDayFormatter(CustomWeekDayFormatter());
+        binding.calendarView.setWeekDayFormatter(calendarDecorator.CustomWeekDayFormatter());
 
         // 연월 표기법 변경
         binding.calendarView.setTitleFormatter { day ->
@@ -66,9 +54,11 @@ class MainActivity : AppCompatActivity() {
         binding.calendarView.setLeftArrow(0)
         binding.calendarView.isPagingEnabled = false
 
-        val saturdayDecorator = SaturdayDecorator(binding.calendarView.context)
-        val sundayDecorator = SundayDecorator(binding.calendarView.context)
-        val selectedMonthDecorator = SelectedMonthDecorator(CalendarDay.today().month)
+
+
+        val saturdayDecorator = calendarDecorator.SaturdayDecorator()
+        val sundayDecorator = calendarDecorator.SundayDecorator()
+        val selectedMonthDecorator = calendarDecorator.SelectedMonthDecorator(CalendarDay.today().month)
 
         // 특정일 밑에 점찍기
 //        val highlightedDate = CalendarDay.from(2024, 1, 24)
@@ -89,11 +79,11 @@ class MainActivity : AppCompatActivity() {
         binding.calendarView.isDynamicHeightEnabled = true
 
         // 날짜가 바뀔때마다 색 지정
-        binding.calendarView.setOnDateChangedListener { widget, date, selected ->
-            val selectedDateDecorator = SelectedDateDecorator(binding.calendarView.context, date)
-            val unSelectedDateDecorator = UnSelectedDateDecorator(binding.calendarView.context, date)
-            val unSelectedSaturdayDecorator = UnSelectedSaturdayDecorator(binding.calendarView.context, date)
-            val unSelectedSundayDecorator = UnSelectedSundayDecorator(binding.calendarView.context, date)
+        binding.calendarView.setOnDateChangedListener { _, date, _ ->
+            val selectedDateDecorator = calendarDecorator.SelectedDateDecorator(date)
+            val unSelectedDateDecorator = calendarDecorator.UnSelectedDateDecorator(date)
+            val unSelectedSaturdayDecorator = calendarDecorator.UnSelectedSaturdayDecorator(date)
+            val unSelectedSundayDecorator = calendarDecorator.UnSelectedSundayDecorator(date)
             binding.calendarView.addDecorators(selectedDateDecorator,unSelectedDateDecorator,unSelectedSaturdayDecorator,unSelectedSundayDecorator,selectedMonthDecorator)
         }
     }
@@ -146,6 +136,12 @@ class MainActivity : AppCompatActivity() {
         isImageChanged = !isImageChanged
     }
 
+    // 버튼 클릭 이벤트 처리
+    fun goToDiaryActivity(view: View) {
+        val intent = Intent(this, DiaryActivity::class.java)
+        startActivity(intent)
+    }
+
     private fun showToast(message: String) {
         // 예시로 Toast 메시지 표시
         Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
@@ -154,137 +150,5 @@ class MainActivity : AppCompatActivity() {
     /**
      * 아래 부터는 Decorator 설정
      */
-    /* 토요일 색상을 설정 */
-    private class SaturdayDecorator(private val context: Context): DayViewDecorator {
-        override fun shouldDecorate(day: CalendarDay): Boolean {
-            val dayOfWeek = day.date.dayOfWeek.value // 1 (월요일) ~ 7 (일요일)
-            return dayOfWeek == 6 // 토요일
-        }
 
-        override fun decorate(view: DayViewFacade) {
-            val color = ContextCompat.getColor(context, R.color.blue)
-            view.addSpan(ForegroundColorSpan(color))
-        }
-    }
-
-    /* 일요일 색상을 설정 */
-    private class SundayDecorator(private val context: Context): DayViewDecorator {
-        override fun shouldDecorate(day: CalendarDay): Boolean {
-            val dayOfWeek = day.date.dayOfWeek.value // 1 (월요일) ~ 7 (일요일)
-            return dayOfWeek == 7 // 일요일
-        }
-
-        override fun decorate(view: DayViewFacade) {
-            val color = ContextCompat.getColor(context, R.color.red)
-            view.addSpan(ForegroundColorSpan(color))
-        }
-    }
-
-    /* 이번달에 속하지 않지만 캘린더에 보여지는 이전달/다음달의 일부 날짜를 설정하는 클래스 */
-    private inner class SelectedMonthDecorator(val selectedMonth : Int) : DayViewDecorator {
-        override fun shouldDecorate(day: CalendarDay): Boolean {
-            return day.month != selectedMonth
-        }
-        override fun decorate(view: DayViewFacade) {
-            view.addSpan(ForegroundColorSpan(ContextCompat.getColor(binding.calendarView.context, R.color.gray50)))
-        }
-    }
-
-    /* 특정 일자 밑에 점을 추가*/
-    private class HighlightDecorator(private val context: Context, private val highlightedDays: Set<CalendarDay>) : DayViewDecorator {
-        override fun shouldDecorate(day: CalendarDay): Boolean {
-            return highlightedDays.contains(day)
-        }
-
-        override fun decorate(view: DayViewFacade) {
-            val color = ContextCompat.getColor(context, R.color.red)
-            view.addSpan(DotSpan(5F, color)) // 날짜에 빨간색 동그라미 추가 (이 부분은 원하는 스타일로 수정 가능)
-        }
-    }
-
-    /* 선택된 일자의 숫자색을 바꿈 */
-    private class SelectedDateDecorator(private val context: Context, private val selectedDate: CalendarDay) : DayViewDecorator {
-        override fun shouldDecorate(day: CalendarDay): Boolean {
-            return day == selectedDate
-        }
-
-        override fun decorate(view: DayViewFacade) {
-            val color = ContextCompat.getColor(context, R.color.white)
-            view.addSpan(ForegroundColorSpan(color))
-        }
-
-    }
-
-    /* 선택되지 않은 일자의 숫자색을 바꿈 */
-    private class UnSelectedDateDecorator(private val context: Context, private val selectedDate: CalendarDay) : DayViewDecorator {
-
-        override fun shouldDecorate(day: CalendarDay): Boolean {
-            return day != selectedDate
-        }
-
-        override fun decorate(view: DayViewFacade) {
-            val color = ContextCompat.getColor(context, R.color.deep)
-            view.addSpan(ForegroundColorSpan(color))
-        }
-    }
-
-    /* 선택되지 않은 토요일의 색을 바꿈 */
-    private class UnSelectedSaturdayDecorator(private val context: Context, private val selectedDate: CalendarDay) : DayViewDecorator {
-        private var dayOfWeek = 0
-        override fun shouldDecorate(day: CalendarDay): Boolean {
-            dayOfWeek = day.date.dayOfWeek.value
-            return day != selectedDate && dayOfWeek == 6
-        }
-
-        override fun decorate(view: DayViewFacade) {
-            val color = ContextCompat.getColor(context, R.color.blue)
-            view.addSpan(ForegroundColorSpan(color))
-        }
-    }
-
-    /* 선택되지 않은 일요일의 색을 바꿈 */
-    private class UnSelectedSundayDecorator(private val context: Context, private val selectedDate: CalendarDay) : DayViewDecorator {
-        private var dayOfWeek = 0
-        override fun shouldDecorate(day: CalendarDay): Boolean {
-            dayOfWeek = day.date.dayOfWeek.value
-            return day != selectedDate && dayOfWeek == 7
-        }
-
-        override fun decorate(view: DayViewFacade) {
-            val color = ContextCompat.getColor(context, R.color.red)
-            view.addSpan(ForegroundColorSpan(color))
-        }
-    }
-
-    /* 요일별로 색을 다르게 */
-    private inner class CustomWeekDayFormatter : WeekDayFormatter {
-        private val saturdayColor = ContextCompat.getColor(binding.calendarView.context, R.color.blue)
-        private val sundayColor = ContextCompat.getColor(binding.calendarView.context, R.color.red)
-
-        override fun format(dayOfWeek: DayOfWeek): CharSequence {
-            val day = dayOfWeek.value
-            val color = when (day) {
-                6 -> saturdayColor // 토요일
-                7 -> sundayColor   // 일요일
-                else -> ContextCompat.getColor(binding.calendarView.context, R.color.deep) // 다른 요일은 검정색
-            }
-
-            return SpannableString("${getDayOfWeekName(day)}").apply {
-                setSpan(ForegroundColorSpan(color), 0, length, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE)
-            }
-        }
-
-        private fun getDayOfWeekName(day: Int): String {
-            return when (day) {
-                1 -> "월"
-                2 -> "화"
-                3 -> "수"
-                4 -> "목"
-                5 -> "금"
-                6 -> "토"
-                7 -> "일"
-                else -> ""
-            }
-        }
-    }
 }
